:truck: api: move certification-details domain builder from lib to src
